requirements.txt
flask==2.0.1
werkzeug==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.6.3
beautifulsoup4==4.9.3
requests==2.26.0